+++ src/auth/auth.module.ts
@@ -3,7 +3,6 @@ import { AuthController } from './auth.controller';
 import { UsersModule } from 'src/users/users.module';
 import { AuthService } from './auth.service';
 import { PassportModule } from '@nestjs/passport';
-import { LocalStrategy } from './local.strategy';
 import { FtOauthStrategy } from './ft-oauth.strategy';
 import { SessionSerializer } from './session.serializer';
 
@@ -15,6 +14,6 @@ import { SessionSerializer } from './session.serializer';
     }),
   ],
   controllers: [AuthController],
-  providers: [AuthService, LocalStrategy, FtOauthStrategy, SessionSerializer],
+  providers: [AuthService, FtOauthStrategy, SessionSerializer],
 })
 export class AuthModule {}
