--- src/common/guards/ft-oauth.guard.ts
@@ -1,10 +1,0 @@
-import { ExecutionContext, Injectable } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-
-@Injectable()
-export class FtOauthGuard extends AuthGuard('42') {
-  async canActivate(context: ExecutionContext) {
-    const request = context.switchToHttp().getRequest();
-    return request.session.userId;
-  }
-}
