--- src/common/guards/login.guard.ts
@@ -1,10 +1,0 @@
-import { ExecutionContext, Injectable } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-
-@Injectable()
-export class LoginGuard extends AuthGuard('local') {
-  async canActivate(context: ExecutionContext) {
-    const request = context.switchToHttp().getRequest();
-    return request.session.userId;
-  }
-}
